+++ controllers/src/test/scala/cromwell/pipeline/controller/AuthControllerTest.scala
@@ -1,14 +1,14 @@
 package cromwell.pipeline.controller
 
 import akka.http.scaladsl.model.ContentTypes.`application/json`
-import akka.http.scaladsl.model.{ HttpEntity, StatusCodes }
+import akka.http.scaladsl.model.{HttpEntity, StatusCodes}
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import cromwell.pipeline.controller.AuthController._
-import cromwell.pipeline.datastorage.dto.auth.{ AuthResponse, SignInRequest, SignUpRequest }
+import cromwell.pipeline.datastorage.dto.formatters.AuthFormatters.{AuthResponse, SignInRequest, SignUpRequest}
 import cromwell.pipeline.datastorage.utils.validator.DomainValidation
 import cromwell.pipeline.service.AuthService
 import org.scalamock.scalatest.MockFactory
-import org.scalatest.{ Assertion, Matchers, WordSpec }
+import org.scalatest.{Assertion, Matchers, WordSpec}
 import play.api.libs.json.Json
 
 import scala.concurrent.Future
