+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dao/entry/ProjectEntry.scala
@@ -2,7 +2,8 @@ package cromwell.pipeline.datastorage.dao
 
 import cromwell.pipeline.datastorage.Profile
 import cromwell.pipeline.datastorage.dao.entry.UserEntry
-import cromwell.pipeline.datastorage.dto.{ Project, ProjectId, UserId }
+import cromwell.pipeline.datastorage.dto.formatters.ProjectFormatters.ProjectId
+import cromwell.pipeline.datastorage.dto.{Project, UserId}
 
 trait ProjectEntry {
   this: Profile with UserEntry =>
