+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dao/repository/ProjectRepository.scala
@@ -2,7 +2,8 @@ package cromwell.pipeline.datastorage.dao.repository
 
 import cromwell.pipeline.database.PipelineDatabaseEngine
 import cromwell.pipeline.datastorage.dao.ProjectEntry
-import cromwell.pipeline.datastorage.dto.{ Project, ProjectId }
+import cromwell.pipeline.datastorage.dto.Project
+import cromwell.pipeline.datastorage.dto.formatters.ProjectFormatters.ProjectId
 
 import scala.concurrent.Future
