+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/User.scala
@@ -1,9 +1,10 @@
 package cromwell.pipeline.datastorage.dto
 
 import cromwell.pipeline.datastorage.dto.User.UserEmail
+import cromwell.pipeline.datastorage.dto.formatters.UserFormatters.ProfilePicture
 import play.api.libs.functional.syntax._
 import play.api.libs.json.Reads._
-import play.api.libs.json.{ Format, Json, OFormat }
+import play.api.libs.json.{Format, Json, OFormat}
 import slick.lifted.MappedTo
 
 final case class User(
@@ -28,8 +29,3 @@ object UserId {
   implicit lazy val userIdFormat: Format[UserId] = implicitly[Format[String]].inmap(UserId.apply, _.value)
 }
 
-final case class ProfilePicture(value: Array[Byte]) extends MappedTo[Array[Byte]]
-
-object ProfilePicture {
-  implicit lazy val profilePictureFormat: OFormat[ProfilePicture] = Json.format[ProfilePicture]
-}
