--- repositories/src/main/scala/cromwell/pipeline/datastorage/dto/auth/AuthResponse.scala
@@ -1,9 +1,0 @@
-package cromwell.pipeline.datastorage.dto.auth
-
-import play.api.libs.json.{ Json, OFormat }
-
-final case class AuthResponse(accessToken: String, refreshToken: String, accessTokenExpiration: Long)
-
-object AuthResponse {
-  implicit val authResponseFormat: OFormat[AuthResponse] = Json.format[AuthResponse]
-}
