--- repositories/src/main/scala/cromwell/pipeline/datastorage/dto/auth/SignUpRequest.scala
@@ -1,9 +1,0 @@
-package cromwell.pipeline.datastorage.dto.auth
-
-import play.api.libs.json.{ Json, OFormat }
-
-final case class SignUpRequest(email: String, password: String, firstName: String, lastName: String)
-
-object SignUpRequest {
-  implicit val signUpRequestFormat: OFormat[SignUpRequest] = Json.format[SignUpRequest]
-}
