--- repositories/src/main/scala/cromwell/pipeline/datastorage/dto/user/UserUpdateRequest.scala
@@ -1,9 +1,0 @@
-package cromwell.pipeline.datastorage.dto.user
-
-import play.api.libs.json.{ Json, OFormat }
-
-final case class UserUpdateRequest(email: String, firstName: String, lastName: String)
-
-object UserUpdateRequest {
-  implicit val updateRequestFormat: OFormat[UserUpdateRequest] = Json.format[UserUpdateRequest]
-}
