+++ repositories/src/main/scala/cromwell/pipeline/datastorage/utils/validator/FormValidatorNel.scala
@@ -3,8 +3,8 @@ package cromwell.pipeline.datastorage.utils.validator
 import cats.data.Validated._
 import cats.data._
 import cats.implicits._
-import cromwell.pipeline.datastorage.dto.auth.SignUpRequest
-import cromwell.pipeline.datastorage.dto.user.PasswordUpdateRequest
+import cromwell.pipeline.datastorage.dto.formatters.AuthFormatters.SignUpRequest
+import cromwell.pipeline.datastorage.dto.formatters.UserFormatters.PasswordUpdateRequest
 
 object FormValidatorNel {
