+++ repositories/src/test/scala/cromwell/pipeline/datastorage/dao/repository/auth/AuthUtilsTest.scala
@@ -2,12 +2,12 @@ package cromwell.pipeline.datastorage.dao.repository.auth
 
 import java.time.Instant
 
-import cromwell.pipeline.datastorage.dto.auth.AuthResponse
-import cromwell.pipeline.datastorage.utils.auth.{ AccessTokenContent, AuthUtils, RefreshTokenContent }
-import cromwell.pipeline.utils.{ AuthConfig, ExpirationTimeInSeconds }
-import org.scalatest.{ Assertion, Matchers, WordSpec }
+import cromwell.pipeline.datastorage.dto.formatters.AuthFormatters.AuthResponse
+import cromwell.pipeline.datastorage.utils.auth.{AccessTokenContent, AuthUtils, RefreshTokenContent}
+import cromwell.pipeline.utils.{AuthConfig, ExpirationTimeInSeconds}
+import org.scalatest.{Assertion, Matchers, WordSpec}
 import pdi.jwt.algorithms.JwtHmacAlgorithm
-import pdi.jwt.{ Jwt, JwtAlgorithm }
+import pdi.jwt.{Jwt, JwtAlgorithm}
 import play.api.libs.json.Json
 
 class AuthUtilsTest extends WordSpec with Matchers {
