+++ repositories/src/test/scala/cromwell/pipeline/datastorage/dao/repository/utils/TestProjectUtils.scala
@@ -2,7 +2,8 @@ package cromwell.pipeline.datastorage.dao.repository.utils
 
 import java.util.UUID
 
-import cromwell.pipeline.datastorage.dto.{ Project, ProjectId, UserId }
+import cromwell.pipeline.datastorage.dto.formatters.ProjectFormatters.ProjectId
+import cromwell.pipeline.datastorage.dto.{Project, UserId}
 
 object TestProjectUtils {
