+++ services/src/main/scala/cromwell/pipeline/service/ProjectService.scala
@@ -5,6 +5,8 @@ import java.util.UUID
 import cromwell.pipeline.datastorage.dao.repository.ProjectRepository
 import cromwell.pipeline.datastorage.dto._
 import cromwell.pipeline.service.Exceptions.{ ProjectAccessDeniedException, ProjectNotFoundException }
+import cromwell.pipeline.datastorage.dto.Project
+import cromwell.pipeline.datastorage.dto.formatters.ProjectFormatters.{ ProjectAdditionRequest, ProjectId }
 
 import scala.concurrent.{ ExecutionContext, Future }
