+++ services/src/main/scala/cromwell/pipeline/service/UserService.scala
@@ -1,11 +1,11 @@
 package cromwell.pipeline.service
 
 import cromwell.pipeline.datastorage.dao.repository.UserRepository
-import cromwell.pipeline.datastorage.dto.user.{ PasswordUpdateRequest, UserUpdateRequest }
-import cromwell.pipeline.datastorage.dto.{ User, UserId, UserNoCredentials }
+import cromwell.pipeline.datastorage.dto.formatters.UserFormatters.{PasswordUpdateRequest, UserNoCredentials, UserUpdateRequest}
+import cromwell.pipeline.datastorage.dto.{User, UserId}
 import cromwell.pipeline.utils.StringUtils
 
-import scala.concurrent.{ ExecutionContext, Future }
+import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Random
 
 class UserService(userRepository: UserRepository)(implicit executionContext: ExecutionContext) {
