+++ services/src/test/scala/cromwell/pipeline/service/AuthServiceTest.scala
@@ -4,13 +4,13 @@ import java.time.Instant
 
 import cromwell.pipeline.datastorage.dao.repository.UserRepository
 import cromwell.pipeline.datastorage.dto.UserId
-import cromwell.pipeline.datastorage.dto.auth.AuthResponse
-import cromwell.pipeline.datastorage.utils.auth.{ AccessTokenContent, AuthContent, AuthUtils, RefreshTokenContent }
-import cromwell.pipeline.utils.{ AuthConfig, ExpirationTimeInSeconds }
+import cromwell.pipeline.datastorage.dto.formatters.AuthFormatters.AuthResponse
+import cromwell.pipeline.datastorage.utils.auth.{AccessTokenContent, AuthContent, AuthUtils, RefreshTokenContent}
+import cromwell.pipeline.utils.{AuthConfig, ExpirationTimeInSeconds}
 import org.scalamock.scalatest.MockFactory
-import org.scalatest.{ Matchers, WordSpec }
+import org.scalatest.{Matchers, WordSpec}
 import pdi.jwt.algorithms.JwtHmacAlgorithm
-import pdi.jwt.{ Jwt, JwtAlgorithm, JwtClaim }
+import pdi.jwt.{Jwt, JwtAlgorithm, JwtClaim}
 import play.api.libs.json.Json
 
 import scala.concurrent.ExecutionContext
