+++ services/src/test/scala/cromwell/pipeline/service/ProjectServiceTest.scala
@@ -3,6 +3,8 @@ package cromwell.pipeline.service
 import cromwell.pipeline.datastorage.dao.repository.ProjectRepository
 import cromwell.pipeline.datastorage.dao.repository.utils.TestProjectUtils
 import cromwell.pipeline.datastorage.dto.{ Project, ProjectAdditionRequest, ProjectId, UserId }
+import cromwell.pipeline.datastorage.dto.Project
+import cromwell.pipeline.datastorage.dto.formatters.ProjectFormatters.ProjectAdditionRequest
 import org.mockito.Matchers.any
 import org.mockito.Mockito.when
 import org.scalatest.{ AsyncWordSpec, BeforeAndAfterAll, Matchers }
