+++ services/src/test/scala/cromwell/pipeline/service/UserServiceTest.scala
@@ -2,11 +2,11 @@ package cromwell.pipeline.service
 
 import cromwell.pipeline.datastorage.dao.repository.UserRepository
 import cromwell.pipeline.datastorage.dao.repository.utils.TestUserUtils
-import cromwell.pipeline.datastorage.dto.user.{ PasswordUpdateRequest, UserUpdateRequest }
-import cromwell.pipeline.datastorage.dto.{ User, UserId, UserNoCredentials }
+import cromwell.pipeline.datastorage.dto.formatters.UserFormatters.{PasswordUpdateRequest, UserNoCredentials, UserUpdateRequest}
+import cromwell.pipeline.datastorage.dto.{User, UserId}
 import cromwell.pipeline.utils.StringUtils._
 import org.mockito.Mockito._
-import org.scalatest.{ AsyncWordSpec, Matchers }
+import org.scalatest.{AsyncWordSpec, Matchers}
 import org.scalatestplus.mockito.MockitoSugar
 
 import scala.concurrent.Future
